server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
beautifulsoup4==4.12.2
    # via -r requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
mysql-connector-repackaged==0.3.1
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
soupsieve==2.4.1
    # via beautifulsoup4
urllib3==2.0.3
    # via requests
